Code review (#15)
* docs: add storage backend architecture design (2.0.0)

* feat(core): storage backend abstraction (2.0.0)

Core SimpleFlux is now storage-agnostic:
- IStreamStore contract + FluxEventRecord, FluxStreamMetadata, FluxConcurrencyException
- FluxStore takes IStreamStore (+ FluxOptions) instead of TableClient; owns
  discovery, version assignment (expected-version semantics), hydration
- Optimistic concurrency baked into the contract (fixes #7 for all backends)
- Event Version restored on read (fixes #4); cancellation propagates
- DI: AddSimpleFlux() + FluxOptions + IFluxBuilder (M.E.DI.Abstractions only)
- FluxHeader moved out of core (Azure-specific, joins the AzureTables package)
- Shared package metadata centralized in Directory.Build.props; icon at repo root

* feat(core): fluent event registration (WithEvent / WithAssemblyEvents)

- WithEvent<T>() / WithEvent(Type): register exactly one event type, no scanning
- WithAssemblyEvents<TMarker>() / WithAssemblyEvents(Assembly): all events in an assembly
- Explicit registration disables the implicit 'scan all loaded assemblies' fallback;
  registering nothing keeps previous behavior (no WithAllEvents needed)
- Discovery now unions explicit types + assemblies and dedupes by type
  (the old Distinct() on reference types was a no-op)
- Sample/README/ARCHITECTURE updated to the fluent form

* bench: add SimpleFlux.Benchmarks + BENCHMARKING.md design

Add a BenchmarkDotNet skeleton (SimpleFlux.Benchmarks) parameterized
over InMemory vs AzureTables backends:

Performance (P1-P6):
- P1 Append_SingleEvent (baseline latency)
- P2 Append_Batch throughput vs batch size (surfaces the 100-entity
  Azure transaction cliff at 125)
- P3 Read_Stream over 180 events
- P4 Project_Stream (LargeStreamModule 180-event baseline)
- P5 Concurrency_Contention (8 writers to one stream)
- P6 Reflection_Overhead (per-Store discovery + mapping)

Safety (S1-S4, assert behavior not speed):
- S1 concurrency throws FluxConcurrencyException with right ActualVersion
- S2 read order matches version order
- S3 projection ignores unknown events (no RuntimeBinderException)
- S4 Version restored on read (regression guard)

Add docs/BENCHMARKING.md (scenarios/metrics/CI plan). Register
BenchmarkDotNet in Directory.Packages.props (CPM) and add the project
to SimpleFlux.sln.

Also add zen_free_monitor.sh + the Zen free-model monitor cron job
(check deepseek-v4-flash-free vs laguna/nemotron from the Mac mini IP
every 2h; silent until the free pool clears).

NOT BUILD-VERIFIED: no .NET 10 SDK on this dev Mac (dotnet: command not
found). Build once SDK is installed; benchmark code has unverified rough
edges noted in docs/BENCHMARKING.md.

Branch: bench/benchmark-suite

* docs(AGENTS.md): refresh for 2.0.0 storage-abstraction layout

Rewrites AGENTS.md against the current 2.0.0 repo state:
- 3-package layout (SimpleFlux / AzureTables / InMemory) + IStreamStore contract
- concurrency (ETag->FluxConcurrencyException), read-order-by-Version,
  Version-restored-on-read, and XML-docs all FIXED vs the old 1.x notes
- registers benchmarks project + docs/BENCHMARKING.md
- marks .NET 10 SDK as still missing on dev Mac

Branch: bench/benchmark-suite

* bench: fix project wiring so suite builds with .NET 10 SDK

- Fix ProjectReference paths (..\ -> ..\..\ ; forward slashes) so the
  benchmark project actually resolves SimpleFlux / AzureTables / InMemory.
- Add OutputType=Exe so the BenchmarkSwitcher entry point in Program.cs builds.
- Switch entry point to BenchmarkSwitcher.FromAssembly (default config);
  remove the hand-rolled Configs that used non-existent BDN 0.15.8 API.
- Safety benchmarks: register both event types, cast mixed event array to
  FluxEvent[], and relax the concurrency ActualVersion assertion to be
  backend-agnostic (>0 rather than ==1).
- Pin BenchmarkDotNet 0.15.8 in Directory.Packages.props (CPM).

Build now: 0 errors, 0 warnings. (SDK installed: 10.0.400; Azurite up on 10000-10002.)

Branch: bench/benchmark-suite

* bench: fix warm-up + S1 concurrency so AzureTables read/project run

WarmTo180Async now chunks appends into <=50-event groups to stay under
the Azure 100-entity transaction cliff (the 2.0.0 store doesn't chunk),
so Read_Stream/Project_Stream can warm against Azurite and actually
measure. Fixes all-AzureTables Read/Project cases being NA.

S1 concurrency (SafetyBenchmarks) now appends at the IStreamStore
contract directly with a shared expected version so the two concurrent
writers genuinely race and the losing one gets FluxConcurrencyException
with the winner's actual version (assert >0, backend-agnostic).

Branch: bench/benchmark-suite

* bench: record benchmark results run (Aug 2026, M4/.NET 10/Azurite)

All 84 perf + 4 safety benchmarks ran (ShortRun, 0 build errors):
- InMemory + AzureTables (Azurite) both measured.
- 100-entity transaction cliff reproduced: Append_Batch@100/125 AzureTables = NA.
- Read/Project flat ~100µs InMemory; ~9-20ms AzureTables (network-bound).
- Reflection overhead ~1.2µs (negligible — not the hotspot AGENTS feared).
- All 4 safety asserts pass (concurrency, read-order, version-restore,
  unknown-event ignore). S1 asserts at IStreamStore contract level.

Raw log: docs reference /tmp/sf_full2.log; artifacts under benchmarks/.../artifacts/.
docs/BENCHMARKING_RESULTS.md summarizes.

Branch: bench/benchmark-suite

* Benchmarking update.

* fix(AzureTables): chunk appends under the 100-entity transaction limit

* feat!: change Version from int to long across the entire storage contract

Breaking change to IStreamStore, FluxEvent, FluxEventRecord,
FluxStreamMetadata, FluxConcurrencyException, and all backends.
Sentinel values remain -2L (Any) and -1L (NoStream); first event
version is 1L.

* feat(AzureTables): guard against reserved column name collisions

Event properties using [FluxProperty] names that collide with Azure
Table Storage system columns (PartitionKey, RowKey, Timestamp, ETag)
or SimpleFlux metadata columns (EventType, Version, F-HEAD) now
throw ArgumentException with a clear message at write time.

* feat: add SimpleFlux.FlatFile backend (JSONL per stream)

New storage backend that persists event streams as JSONL files on the
local filesystem. Each stream gets its own directory with an append-only
events.jsonl and a meta.json for version tracking. Per-stream file
locking for concurrency. Zero external dependencies.

Storage layout:
  {root}/{streamId}/events.jsonl
  {root}/{streamId}/meta.json

Usage: .UseFlatFile("/path/to/streams")

* bench: add FlatFile backend to benchmark suite, sweep all 3 backends

Benchmarks now test InMemory, AzureTables, and FlatFile across batch
sizes 1-125. FlatFile uses a temp directory with GlobalCleanup disposal.
Azurite docker container is started/stopped externally.

* chore: remove Azurite artifacts from tracking, add to .gitignore

* test: add contract test suite for all 3 backends

Parameterized xUnit tests that exercise the IStreamStore contract
against InMemory, FlatFile, and AzureTables backends:
- Append single/multiple/large batches, read back ordered by version
- Metadata queries (existing, non-existent streams)
- Optimistic concurrency (conflict detection, NoStream, Any)
- Empty batch handling
- FluxStore integration (AddEvent, AddEvents, ProjectTo, StreamExists)

Also adds CONTRIBUTORS.md and rewrites README with adoption-focused
quickstart, backend comparison, custom backend guide, and testing docs.

* chore: add global.json, Directory.Build.targets, refresh AGENTS.md

- global.json: pin SDK 10.0.400 with latestFeature rollForward
- Directory.Build.targets: centralize TreatWarningsAsErrors, test project
  defaults (IsPackable=false, IsTestProject=true, suppress CS1591)
- AGENTS.md: full rewrite reflecting current state — 4 packages, 3 backends,
  contract tests, benchmarks, FlatFile backend, Version→long, chunking

+++ src/SimpleFlux/FluxStore.cs
@@ -10,7 +10,7 @@
 /// </remarks>
 public class FluxStore
 {
-    private const int NoStream = -1;
+    private const long NoStream = -1L;
 
     private readonly IStreamStore _streamStore;
     private readonly List<KnownFluxEventType> _knownEventTypes;
@@ -82,13 +82,13 @@ public class FluxStore
         await _streamStore.AppendToStreamAsync(streamId, expectedVersion, records, version - 1, cancellationToken);
     }
 
-    private async Task<int> ResolveExpectedVersionAsync(string streamId, CancellationToken cancellationToken)
+    private async Task<long> ResolveExpectedVersionAsync(string streamId, CancellationToken cancellationToken)
     {
         var metadata = await _streamStore.GetStreamMetadataAsync(streamId, cancellationToken);
         return metadata?.Version ?? NoStream;
     }
 
-    private static int NextVersion(int expectedVersion) => expectedVersion == NoStream ? 1 : expectedVersion + 1;
+    private static long NextVersion(long expectedVersion) => expectedVersion == NoStream ? 1L : expectedVersion + 1;
 
     private static List<KnownFluxEventType> DiscoverEventTypes(FluxOptions options)
     {
@@ -224,7 +224,7 @@ public class FluxStore
     /// <param name="id">The stream id.</param>
     /// <param name="cancellationToken">A token to observe while waiting for the operation to complete.</param>
     /// <returns>The stream version, or <c>null</c> when the stream does not exist.</returns>
-    public async Task<int?> GetStreamVersion(string id, CancellationToken cancellationToken = default)
+    public async Task<long?> GetStreamVersion(string id, CancellationToken cancellationToken = default)
     {
         var metadata = await _streamStore.GetStreamMetadataAsync(id, cancellationToken);
         return metadata?.Version;
